Attempting to add debug logs for ENQUEUING an invalid object
We were seeing GC Corruption when _popping_ an invalid object, but we
don't know how we got this invalid pointer. (And we can't repro on linux
so can't use RR.)

This is an attempt to check for the object's validity _before enqueuing_
so that we can hopefully give a more useful error message (which
object's pointer was corrupted).

But for some reason it's currently complaining about the very first push....